refact: moving reservation files to reservation pages folder
