+++ frontend/src/app/reservation/pages/EditPay/index.tsx
@@ -1,22 +1,35 @@
-import React, { useState } from 'react';
-import { JustSpider } from '../components/just-spider';
-import { Box, Text, Stack, Icon, HStack, FormControl, FormLabel, Button, ButtonGroup } from '@chakra-ui/react';
-import { FaArrowLeft, FaCheck, FaWifi, FaCar, FaCoffee, FaSnowflake, FaConciergeBell } from 'react-icons/fa';
+import React, { useState} from 'react';
+import {JustSpider} from '../../components/just-spider'
+import { JustNet } from '../../components/just-net';
 import Select, { SingleValue } from 'react-select';
-import { NavBar } from '../../../shared/components/nav-bar';
+import {
+    Box,
+    Text,
+    Button,
+    Icon,
+    ButtonGroup,
+    Stack,
+    FormControl,
+    FormLabel,
+  } from '@chakra-ui/react';
+import {
+    FaCheck,
+    FaArrowLeft,
+} from 'react-icons/fa';
+import { NavBar } from '../../../../shared/components/nav-bar';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 interface OptionType {
-  value: string;
-  label: string;
+    value: string;
+    label: string;
 }
 
-const PayReservation: React.FC = () => {
+const EditPay: React.FC = () => {
   const [selectedPayment, setSelectedPayment] = useState<SingleValue<OptionType>>(null);
 
   const goBack = async () => {
-    window.location.href = "http://localhost:3000/create-reservation";
+    window.location.href = "http://localhost:3000/edit-reservation";
   };
 
   const handlePayment = async () => {
@@ -24,7 +37,7 @@ const PayReservation: React.FC = () => {
       toast.warning("Preencha todos os campos!");
       return;
     }
-    toast.success("Reserva realizada com sucesso!", {
+    toast.success("Reserva atualizada com sucesso!", {
       autoClose: 3000, // 3000ms = 3 seconds
       onClose: () => {
         window.location.href = "http://localhost:3000/my-reservations";
@@ -67,7 +80,7 @@ const PayReservation: React.FC = () => {
     })
   };
 
-  return (
+    return (
     <Box width="100vw" height="100vh" display="flex" flexDirection="column">
       <NavBar />
       <Box
@@ -78,71 +91,8 @@ const PayReservation: React.FC = () => {
         alignItems="center"
         position="relative"
       >
-         <JustSpider />
-        {/* Bloco roxo com sombra */}
-        <Box position="absolute" left="200px" top="100px" width="500px">
-          <Box
-            position="relative"
-            width="100%"
-            height="300px"
-            bg="#6A0572"
-            zIndex="1"
-          />
-          <Box
-            position="absolute"
-            width="100%"
-            height="300px"
-            bg="rgba(255, 255, 255, 0.1)"
-            top="10px"
-            left="10px"
-            zIndex="0"
-          />
-          <Text
-            mt="20px"
-            color="#EAEAEA"
-            fontSize="2xl"
-            fontWeight="bold"
-            textAlign="center"
-          >
-            Zumbi Digital
-          </Text>
-          <Text color="#EAEAEA" fontSize="xl" textAlign="center">
-            R$ 1200,00 a diária
-          </Text>
-          <HStack mt="10px" justify="center" spacing={4}>
-            <HStack spacing={1}>
-              <Icon as={FaWifi} color="#EAEAEA" />
-              <Text color="#EAEAEA" fontSize="sm" whiteSpace="nowrap">
-                Wi-Fi
-              </Text>
-            </HStack>
-            <HStack spacing={1}>
-              <Icon as={FaConciergeBell} color="#EAEAEA" />
-              <Text color="#EAEAEA" fontSize="sm" whiteSpace="nowrap">
-                Serviço de Quarto
-              </Text>
-            </HStack>
-            <HStack spacing={1}>
-              <Icon as={FaCoffee} color="#EAEAEA" />
-              <Text color="#EAEAEA" fontSize="sm" whiteSpace="nowrap">
-                Café da Manhã
-              </Text>
-            </HStack>
-            <HStack spacing={1}>
-              <Icon as={FaSnowflake} color="#EAEAEA" />
-              <Text color="#EAEAEA" fontSize="sm" whiteSpace="nowrap">
-                Ar-condicionado
-              </Text>
-            </HStack>
-            <HStack spacing={1}>
-              <Icon as={FaCar} color="#EAEAEA" />
-              <Text color="#EAEAEA" fontSize="sm" whiteSpace="nowrap">
-                Estacionamento
-              </Text>
-            </HStack>
-          </HStack>
-        </Box>
-
+        <JustSpider/>
+        <JustNet/>
         <Box
           position="relative"
           width="400px"
@@ -152,22 +102,22 @@ const PayReservation: React.FC = () => {
           flexDirection="column"
           justifyContent="center"
           alignItems="center"
-          left="300px"
+          left="0px" // Ajuste a posição para a esquerda
           py="20px"
           top="0"
         >
-          <Text
-            position="absolute"
-            top="-10px"
-            bg="#191919"
-            px="10px"
-            color="#EAEAEA"
-          >
-            Pagamento
-          </Text>
-          <Stack spacing={4} mt="30px" align="center" width="80%">
+            <Text
+                position="absolute"
+                top="-10px"
+                bg="#191919"
+                px="10px"
+                color="#EAEAEA"
+            >
+            Pagamento Reserva
+            </Text>
+            <Stack spacing={4} mt="30px" align="center" width="80%">
             <Text color="#EAEAEA" fontSize="2xl" textAlign="center">
-              Valor total: R$ 3600,00
+              Novo valor total: R$ 3600,00
             </Text>
             <FormControl id="paymentMethod" isRequired>
               <FormLabel color="#EAEAEA">Método de Pagamento</FormLabel>
@@ -215,9 +165,11 @@ const PayReservation: React.FC = () => {
           </ButtonGroup>
         </Box>
       </Box>
-      <ToastContainer theme="dark" />
+      <ToastContainer theme="dark"/>
     </Box>
-  );
-};
 
-export default PayReservation;
+
+    );
+}
+
+export default EditPay;
