+++ frontend/src/app/reservation/pages/EditReservation/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState, ForwardedRef } from 'react';
-import {JustSpider} from '../components/just-spider'
-import { JustNet } from '../components/just-net';
+import {JustSpider} from '../../components/just-spider'
+import { JustNet } from '../../components/just-net';
 import {
     Box,
     Text,
@@ -27,7 +27,7 @@ import {
   } from 'react-icons/fa';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import { NavBar } from '../../../shared/components/nav-bar';
+import { NavBar } from '../../../../shared/components/nav-bar';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
